Implements  background GameObjects
This comes in contrast with bacground being just
repeated images: if there is a GameObject class
with the same name a color has in the palette,
it is instantiated, instead of a simple image load
from a file with the same name.

This "background" object can have a hardness attribute
which obstructs Actors passage, and can react to being
walked over by Actors (Touched).

TODO: make reactions available on actuall touch
(not just walking over).